Added routes and moved default app data to Home component
